requirements.txt
streamlit
pytube
Pillow
requests
